requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncer==0.0.2
attrs==23.2.0
awscli==1.32.115
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
bidict==0.23.1
bioimageio.core==0.6.3
bioimageio.spec==0.5.2.post4
bleach==6.1.0
blinker==1.8.2
bm3d==4.0.1
bm4d==4.2.3
botocore==1.34.115
build==1.1.1
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
chevron==0.14.0
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.0
csbdeep==0.8.0
cycler==0.10.0
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dm-tree==0.1.8
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
etils==1.8.0
executing==2.0.1
fastapi==0.108.0
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
filelock==3.13.3
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.1
humanfriendly==10.0
idna==2.10
imagecodecs==2024.1.1
imageio==2.34.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==6.11.0
importlib_resources==6.4.0
imutils==0.5.4
ipykernel==6.29.4
ipython==8.23.0
ipython-extensions==0.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kagglehub==0.2.5
keras==3.5.0
keras-core==0.1.7
keras-cv==0.9.0
keras-self-attention==0.51.0
kiwisolver==1.4.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.34
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.33
Lazify==0.4.0
lazy_loader==0.4
libclang==18.1.1
literalai==0.0.300
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.5.1
metakernel==0.30.2
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
n2v==0.3.3
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
oauthlib==3.2.2
octave_kernel==0.36.0
onnxruntime==1.17.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.9.15
overrides==7.7.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
plotly==5.20.0
plotly-express==0.4.1
pooch==1.8.1
posthog==3.5.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==25.1.2
rawpy==0.21.0
referencing==0.34.0
regex==2023.12.25
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.30
rpds-py==0.18.0
rsa==4.7.2
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==68.2.2
silence-tensorflow==1.2.1
simple-websocket==1.0.0
SimpleITK==2.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.32.0.post1
statsmodels==0.14.1
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.3
tensorflow-metadata==1.14.0
tensorrt==10.4.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.5.3
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.11.0
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.196
uptrace==1.22.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xarray==2024.3.0
yarl==1.9.4
zipp==3.18.1

